MACVLAN: better error handling for unrealized parent interface
When creating a MACVLAN device it can have an unrealized parent. In that
case, the parent won't have an ifindex and it will cause a crash due to
an assert on parent being missing.

In order to avoid this, we are handling the situation similar to what we
do for VLAN devices. NetworkManager will raise different error and block
the autoconnection instead of crashing.

This solves the crash for the following commands:

```
nmcli connection add type macvlan ifname mv1 con-name mv1+ macvlan.parent dummy0 mode vepa
nmcli connection add type dummy ifname dummy0 con-name dummy0+ autoconnect no
```