build: fix pretty() sed invocation portability in .ver script
The `\0` being used to refer to the whole match is a GNU
extension. On other implementations of sed, this will simply
replace the each matched line with `	0;`, resulting in symbols
being missing in the NetworkManager executable, which breaks the
wifi plugin and others.

Use `&` instead, which behaves as expected in all implementations.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1562