glib-aux: make NMIPAddr a plain union
It is slightly confusing to be required to be aware whether something is
a union or a struct. Hence, the union was wrapped in a struct.

However, we anyway almost always use the typedef NMIPAddr.  The single
place where we forward declare the type, we can correctly use the union
specifier.