glib-aux/prioq: ensure to clear "idx" when removing item from NMPrioq
NMPrioq is copied from systemd's "prioq.c". It uses the index pointer to
find elements in the heap directly. However, it's not very careful about
clearing or maintaining the index pointer if the element is not in the
queue. Like, find_item() accepts a bogus index and only finds the data
if the data and the index matches. That seems error prone.

In the first step, when we remove an item from the queue, ensure to
reset the index to NM_PRIOQ_IDX_NULL.