triage: only make stale issues/merge-requests with a label instead of autoclosing them
An issue/merge-requests only moves forward by a human putting effort in.
Wether the issue is open, closed, autoclosed or flagged with a "stale"
label doesn't matter in theory.

In practice, humans don't like getting their issues/merge-requests
autoclosed. Instead, add a "stale" label.

There was always a benefit of this automatic action. It tended to
generate some feedback and new action. Sometimes it even brought new
attention to the issue/merge-request to get it fixed. That is hopefully
also the case with the stale label.